Re-organize parser code and split common structs to their own module
