+++ src/main.rs
@@ -1,5 +1,3 @@
-mod parser;
-
 use std::path::PathBuf;
 
 use structopt::StructOpt;
